requirements.txt
streamlit
openai
python-docx
pytesseract
pillow
pandas
openpyxl
PyMuPDF